+++ app/frontend/src/pages/lights/LightsPage.svelte
@@ -2,12 +2,13 @@
   import { Stack } from '@naco-ui/svelte'
   import { fsd } from 'feature-sliced-svelte'
 
-  import { LightsDomainSettings } from '$widgets/lights'
+  import { LightsDomainSettings, LightsPreset } from '$widgets/lights'
 </script>
 <div use:fsd={'pages/LightsPage'}>
   <Stack gap="m">
     <LightsDomainSettings domain="backlight" />
     <LightsDomainSettings domain="halo" />
     <LightsDomainSettings domain="sidelight" />
+    <LightsPreset />
   </Stack>
 </div>