+++ app/internal/errors/errors.go
@@ -1,0 +1,7 @@
+package errors
+
+import "errors"
+
+var (
+	ErrPresetWrongDevice = errors.New("PRESET_WRONG_DEVICE")
+)