+++ app/internal/repository/repository.go
@@ -5,6 +5,7 @@ import (
 	"nuga_ui/internal/interfaces"
 	"nuga_ui/internal/repository/device"
 	"nuga_ui/internal/repository/environment"
+	"nuga_ui/internal/repository/preset"
 	"nuga_ui/internal/repository/settings"
 )
 
@@ -18,5 +19,6 @@ func New(filePath string) (*interfaces.Repository, error) {
 		Device:      device.New(),
 		Settings:    settings,
 		Environment: environment.New(),
+		Preset:      preset.New(),
 	}, nil
 }